Added logic of PositionType and logic to createa and update workers

+++ MedicalExamination.BLL/PositionTypeService.cs
@@ -1,48 +1,123 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using MedicalExamination.DAL;
 using MedicalExamination.Entities;
+using SimpleMapper;
 
 namespace MedicalExamination.BLL
 {
     public sealed class PositionTypeService : IPositionTypeService
     {
         private readonly IGenericRepository<PositionType> _positionTypeRepository;
+        private readonly IGenericRepository<ProvideService> _provideServiceRepository;
+        private readonly IGenericRepository<ServiceType> _serviceTypeRepository;
+        private readonly IProvideServiceService _provideServiceService;
+        private readonly IPositionService _positionService;
 
-        public PositionTypeService(IGenericRepository<PositionType> positionTypeRepository)
+        public PositionTypeService(IGenericRepository<PositionType> positionTypeRepository,
+            IGenericRepository<ProvideService> provideServiceRepository,
+            IGenericRepository<ServiceType> serviceTypeRepository, IProvideServiceService provideServiceService, IPositionService positionService)
         {
             _positionTypeRepository = positionTypeRepository;
+            _provideServiceRepository = provideServiceRepository;
+            _serviceTypeRepository = serviceTypeRepository;
+            _provideServiceService = provideServiceService;
+            _positionService = positionService;
         }
 
-        public IEnumerable<PositionType> GetAllPositionTypes()
+        public IEnumerable<PositionTypeModel> GetAllPositionTypes()
         {
-            return _positionTypeRepository.GetAll();
+            var positionTypes = _positionTypeRepository.GetAll().AsEnumerable().Map<PositionType, PositionTypeModel>();
+
+            return positionTypes;
         }
 
-        public PositionType GetPositionType(Guid id)
+        public PositionTypeModel GetPositionType(Guid id)
         {
-            return _positionTypeRepository.GetById(id);
+            var position =
+                SimpleMapper.Mapper.Map<PositionType, PositionTypeModel>(_positionTypeRepository.GetAll()
+                    .FirstOrDefault(pt => pt.Id == id));
+            position.ProvideServices = _provideServiceRepository.GetAll().Where(ps => ps.PositionId == position.Id).AsEnumerable()
+                .Map<ProvideService, ProvideServiceModel>();
+
+            var serviceTypes = _serviceTypeRepository.GetAll().ToArray();
+            position.ServiceTypes = serviceTypes.Where(st => position.ProvideServices.Any(ps => ps.ServiceId == st.Id))
+                .Map<ServiceType, ServiceTypeModel>();
+
+            return position;
+        }
+
+        public IEnumerable<PositionTypeModel> GetPositionTypesByPositionIds(IEnumerable<Guid> positionIds)
+        {
+            var positionTypes = _positionTypeRepository.GetAll();
+            positionTypes = positionTypes.Where(p => positionIds.Any(id => id == p.Id));
+
+            return positionTypes.AsEnumerable().Map<PositionType, PositionTypeModel>();
         }
 
         public void CreatePositionType(PositionTypeModel positionTypeModel)
         {
             var positionType = SimpleMapper.Mapper.Map<PositionTypeModel, PositionType>(positionTypeModel);
 
-            _positionTypeRepository.Insert(positionType);
+            var createdPositionType = _positionTypeRepository.Insert(positionType);
+
+            var provideServices = positionTypeModel.ServiceTypes.Select(serviceType =>
+                new ProvideService {PositionId = createdPositionType.Id, ServiceId = serviceType.Id}).ToList();
+
+            _provideServiceRepository.Insert(provideServices);
         }
 
         public void UpdatePositionType(PositionTypeModel positionTypeModel)
         {
             var positionType = SimpleMapper.Mapper.Map<PositionTypeModel, PositionType>(positionTypeModel);
-
             _positionTypeRepository.Update(positionType);
+
+            var existingProvideServices = _provideServiceService.GetProvideServicesOfPosition(positionTypeModel.Id);
+
+            var notExistingProvideServices =
+                positionTypeModel.ServiceTypes.Where(st => existingProvideServices.All(pd => pd.ServiceId != st.Id));
+
+            var newProvideServices = notExistingProvideServices.Select(serviceType =>
+                new ProvideService {PositionId = positionTypeModel.Id, ServiceId = serviceType.Id}).ToList();
+
+            if (newProvideServices.Count > 0)
+            {
+                _provideServiceRepository.Insert(newProvideServices);
+            }
+
+            var unselectedExistingProvideService = existingProvideServices.Where(eps =>
+                    positionTypeModel.ServiceTypes.Any(st => st.Id != eps.ServiceId))
+                .Select(ps => new ProvideService {Id = ps.Id, PositionId = ps.PositionId, ServiceId = ps.ServiceId}).ToArray();
+
+            foreach (var provideService in unselectedExistingProvideService)
+            {
+                provideService.Position = _positionTypeRepository.GetById(provideService.PositionId);
+                provideService.Service = _serviceTypeRepository.GetById(provideService.ServiceId);
+            }
+
+            if (unselectedExistingProvideService.Length > 0)
+            {
+                _provideServiceRepository.Delete(unselectedExistingProvideService);
+            }
         }
 
         public void DeletePositionType(PositionTypeModel positionTypeModel)
         {
             var positionType = SimpleMapper.Mapper.Map<PositionTypeModel, PositionType>(positionTypeModel);
 
+            var workerInThisPosition = _positionService.GetWorkersOfSpecifiedPosition(positionTypeModel.Id).ToArray();
+
+            if (workerInThisPosition.Length > 0)
+            {
+                throw new Exception("Данная позиция уже используется врачами. Замените позиции и потом пробуйте снова.");
+            }
+
+            var provideServices = _provideServiceService.GetProvideServicesOfPosition(positionTypeModel.Id);
+
             _positionTypeRepository.Delete(positionType);
+
+            _provideServiceService.DeleteProvideServices(provideServices);
         }
     }
 }
